# src/Resources/Schema/StoreApi/openapi.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/paypal/express/prepare-checkout": {
            "post": {
                "tags": [
                    "Store API",
                    "PayPal"
                ],
                "description": "Logs in a guest customer, with the data of a paypal order",
                "operationId": "preparePayPalExpressCheckout",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "token": {
                                        "description": "ID of the paypal order",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "The url to redirect to",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "redirectUrl": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/paypal/pui/payment-instructions/{transactionId}": {
            "get": {
                "tags": [
                    "Store API",
                    "PayPal"
                ],
                "description": "Tries to get payment instructions for PUI payments",
                "operationId": "getPUIPaymentInstructions",
                "parameters": [
                    {
                        "name": "transactionId",
                        "in": "path",
                        "description": "Identifier of the order transaction to be fetched",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "pattern": "^[0-9a-f]{32}$"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The payment instructions of the order"
                    }
                }
            }
        },
        "/paypal/vault/clear": {
            "post": {
                "tags": [
                    "Store API",
                    "PayPal"
                ],
                "description": "Clears the vault for the current customer",
                "operationId": "paypalVaultClear",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "cancel",
                                            "browser",
                                            "error"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Vault has been cleared successfully"
                    }
                }
            }
        },
...
        "/paypal/create-order": {
            "post": {
                "tags": [
                    "Store API",
                    "PayPal"
                ],
                "description": "Creates a PayPal order from the existing cart or an order",
                "operationId": "createPayPalOrder",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "product": {
                                        "description": "Use an existing order id to create PayPal order",
                                        "type": "string",
                                        "default": "ppcp"
                                    },
                                    "orderId": {
                                        "description": "Use an existing order id to create PayPal order",
                                        "type": "string"
                                    },
                                    "returnUrl": {
                                        "description": "Return URL for the order experience context",
                                        "type": "string"
                                    },
                                    "cancelUrl": {
                                        "description": "Cancel URL for the order experience context",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Returns the created PayPal order id",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/paypal/payment-method-eligibility": {
            "post": {
                "tags": [
                    "Store API",
                    "PayPal"
                ],
                "description": "Sets ineligible payment methods to be removed from the current sales channel context",
                "operationId": "setPaymentMethodEligibility",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "paymentMethods": {
                                        "description": "List of PayPal payment method identifiers according to constant REMOVABLE_PAYMENT_HANDLERS",
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Success"
                    }
                }
            }
        }
    },
...
}